style: renamed `hasSlottedChildren` to `has-slotted-children`
